+++ apps/graphql-auth/src/config/auth/jwt.depedency.ts
@@ -1,0 +1,6 @@
+import { JwtModule } from '@nestjs/jwt';
+
+export const jwtDepedency = JwtModule.register({
+  secret: process.env.JWT_SECRET,
+  signOptions: { expiresIn: '48h' },
+});
