+++ apps/graphql-auth/src/domain/interfaces/IContact.interface.ts
