+++ apps/graphql-auth/src/domain/interfaces/IUsecase.interface.ts
