+++ apps/graphql-auth/src/infra/amqp/rabbitMQ.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { rabbitMQDependency } from './depedencie';
+import { RabbitMQService } from './services/rabbitmq.service';
+
+@Module({
+  imports: [...rabbitMQDependency],
+  exports: [...rabbitMQDependency],
+  providers: [RabbitMQService],
+})
+export class RabbitmqModule {}
