+++ apps/graphql-auth/src/infra/amqp/services/rabbitmq.service.ts
@@ -1,9 +1,12 @@
 import { AmqpConnection } from '@golevelup/nestjs-rabbitmq';
-import { Injectable } from '@nestjs/common';
+import { Inject, Injectable } from '@nestjs/common';
 
 @Injectable()
 export class RabbitMQService {
-  constructor(private readonly amqpConnection: AmqpConnection) {}
+  constructor(
+    @Inject(AmqpConnection)
+    private readonly amqpConnection: AmqpConnection,
+  ) {}
   public async send(pattern: string, data: any) {
     return this.amqpConnection.publish(process.env.EXCHANGE, pattern, data);
   }
