+++ apps/graphql-auth/src/presetation/graphql/depedencies/graphql.depedency.ts
@@ -1,0 +1,8 @@
+import { ApolloDriver, ApolloDriverConfig } from '@nestjs/apollo';
+import { GraphQLModule } from '@nestjs/graphql';
+import { join } from 'path';
+
+export const graphqlDepedency = GraphQLModule.forRoot<ApolloDriverConfig>({
+  driver: ApolloDriver,
+  autoSchemaFile: join(process.cwd(), '/apps/graphql-auth/src/schema.gql'),
+});
