+++ tsconfig.json
@@ -17,6 +17,10 @@
     "strictBindCallApply": false,
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
-    "paths": {}
+    "paths": {
+      // "@src/*": [
+      //   "graphql-auth/src/*"
+      // ]
+    }
   }
 }
